Convert to docker-compose, package rebrow, add blueprint model
