+++ Dockerfile
@@ -4,9 +4,13 @@ MAINTAINER Jonathan Kelley <jonk@omg.lol>
 
 ENV DEBIAN_FRONTEND noninteractive
 
-ADD requirements.txt /
-RUN pip install -r /requirements.txt
+#ADD requirements.txt /
+#RUN pip install -r /requirements.txt
 ADD . /app/
 
+WORKDIR /app
+
+RUN python3 setup.py install
+
 EXPOSE 5001
-ENTRYPOINT ["python", "-u", "/app/runserver.py"]
+ENTRYPOINT ["rebrow"]
