+++ rebrow/data/appendonly.aof
@@ -1,0 +1,84 @@
+*2
+$6
+SELECT
+$1
+0
+*3
+$3
+SET
+$5
+mykey
+$5
+Hello
+*3
+$3
+SET
+$10
+anotherkey
+$23
+will expire in a minute
+*3
+$9
+PEXPIREAT
+$10
+anotherkey
+$13
+1612588417884
+*2
+$6
+SELECT
+$1
+0
+*2
+$3
+DEL
+$10
+anotherkey
+*2
+$6
+SELECT
+$1
+0
+*3
+$3
+SET
+$10
+anotherkey
+$23
+will expire in a minute
+*3
+$9
+PEXPIREAT
+$10
+anotherkey
+$13
+1612588545437
+*2
+$3
+DEL
+$10
+anotherkey
+*2
+$6
+SELECT
+$1
+0
+*3
+$3
+SET
+$5
+mykey
+$10
+{"S": "S"}
+*2
+$6
+SELECT
+$1
+0
+*3
+$3
+SET
+$5
+mykey
+$10
+{"S": "S"}
