+++ setup.cfg
@@ -1,0 +1,5 @@
+[install]
+optimize = 1
+
+[easy_install]
+zip_ok = False
